requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
idna==3.7
python-telegram-bot==13.15
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tornado==6.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wikipedia==1.4.0
